Refactor data module to interop; add conversion utilities
