+++ src/lib.rs
@@ -1,4 +1,4 @@
-mod data;
+mod interop;
 mod ops;
 mod py;
